// src/pages/Agents/Agents.interfaces.ts
export interface Agent {
  uuid: string;
  displayName: string;
  description: string;
  developerName: string;
  characterTags: string[] | null;
  displayIcon: string;
  displayIconSmall: string;
  bustPortrait: string;
  fullPortrait: string;
  fullPortraitV2: string;
  killfeedPortrait: string;
  background: string;
  backgroundGradientColors: string[];
  assetPath: string;
  abilities: Ability[];
}

export interface Ability {
  displayName: string;
  description: string;
  displayIcon: string;
}